feat: Implement complete authentication and order management system
## Backend Enhancements
- Add /auth/me endpoint for current user info with role-based permissions
- Integrate JWT authentication into OrderController
- Extract user ID from JWT token for order operations
- Fix all compilation errors (enum issues, field naming, type mismatches)

## Authentication Features
- Implement comprehensive permission system by user role:
  * ADMIN: Full system access
  * CORPORATE: Company-wide order management
  * PARTNER: Partner order visibility and commission tracking
  * GENERAL: Basic order creation and viewing
- Auto-attach JWT token to all API requests
- Handle 401 errors with automatic login redirect

## Order Management
- Complete order creation flow with business rules:
  * Auto CBM calculation from dimensions
  * Shipping type auto-switch when CBM > 29m³
  * THB 1,500 threshold warnings
  * Member code validation
- Add helper method to extract user from JWT in controllers
- Fix OrderBox entity field mappings (width, height, depth)
- Fix BigDecimal quantity handling in CBM calculations

## Technical Improvements
- Replace switch statements with if-else for enum compatibility
- Fix OrderService box creation with proper field names
- Correct BigDecimal operations in OrderBusinessRuleService
- Ensure all entities use proper data types

Successfully tested compilation and Spring Boot startup with H2 database.

+++ JJYCS/backend/src/main/java/com/ycs/lms/service/OrderBusinessRuleService.java
@@ -123,7 +123,7 @@ public class OrderBusinessRuleService {
                 BigDecimal itemCbm = item.getWidth()
                     .multiply(item.getHeight())
                     .multiply(item.getDepth())
-                    .multiply(BigDecimal.valueOf(item.getQuantity()))
+                    .multiply(item.getQuantity())
                     .divide(BigDecimal.valueOf(1000000), 6, RoundingMode.HALF_UP);
                 
                 totalCbm = totalCbm.add(itemCbm);
